Blasr: refactor, no api change, move functions to headers.
