+++ include/RegisterFilterOptions.h
@@ -1,4 +1,4 @@
-#include "libconfig.h" 
+#include "libconfig.h"
 #include "CommandLineParser.hpp"
 #include "datastructures/alignment/FilterCriteria.hpp"
 #include <sstream>
@@ -6,31 +6,31 @@ using namespace std;
 
 /// Register options for filtering alignments.
 void RegisterFilterOptions(CommandLineParser & clp, int & minAlnLength,
-                           float & minPctSimilarity, float & minPctAccuracy, 
+                           float & minPctSimilarity, float & minPctAccuracy,
                            string & hitPolicyStr, bool & useScoreCutoff,
                            int & scoreSignInt, int & scoreCutoff) {
     ScoreSign ss = static_cast<ScoreSign>(scoreSignInt);
     Score sc(static_cast<float>(scoreCutoff),  ss);
-    FilterCriteria fc(static_cast<DNALength>(minAlnLength), 
-                      minPctSimilarity, minPctAccuracy, 
+    FilterCriteria fc(static_cast<DNALength>(minAlnLength),
+                      minPctSimilarity, minPctAccuracy,
                       useScoreCutoff, sc);
 
     HitPolicy hp("randombest", ScoreSign::NEGATIVE);
 
-    clp.RegisterIntOption("minAlnLength", &minAlnLength, 
+    clp.RegisterIntOption("minAlnLength", &minAlnLength,
                           fc.MinAlnLengthHelp(),
                           CommandLineParser::PositiveInteger);
-    clp.RegisterIntOption("minAlignLength", &minAlnLength, 
-                          "Alias of -minAlnLength", 
+    clp.RegisterIntOption("minAlignLength", &minAlnLength,
+                          "Alias of -minAlnLength",
                           CommandLineParser::PositiveInteger);
-    clp.RegisterIntOption("minLength", &minAlnLength, 
-                          "Alias of -minAlnLength", 
+    clp.RegisterIntOption("minLength", &minAlnLength,
+                          "Alias of -minAlnLength",
                           CommandLineParser::PositiveInteger);
 
-    clp.RegisterFloatOption("minPctSimilarity", &minPctSimilarity, 
-                            fc.MinPctSimilarityHelp(), 
+    clp.RegisterFloatOption("minPctSimilarity", &minPctSimilarity,
+                            fc.MinPctSimilarityHelp(),
                             CommandLineParser::PositiveFloat);
-    clp.RegisterFloatOption("minPctIdentity", &minPctSimilarity, 
+    clp.RegisterFloatOption("minPctIdentity", &minPctSimilarity,
                             "Alias of -minPctSimilarity",
                             CommandLineParser::PositiveFloat);
 
@@ -42,12 +42,12 @@ void RegisterFilterOptions(CommandLineParser & clp, int & minAlnLength,
                             CommandLineParser::PositiveFloat);
 
     clp.RegisterStringOption("hitPolicy", &hitPolicyStr, hp.Help());
-            
-    clp.RegisterIntOption("scoreSign", &scoreSignInt, 
+
+    clp.RegisterIntOption("scoreSign", &scoreSignInt,
                           fc.ScoreSignHelp(),
                           CommandLineParser::Integer);
 
-    clp.RegisterIntOption("scoreCutoff", &scoreCutoff, 
+    clp.RegisterIntOption("scoreCutoff", &scoreCutoff,
                           fc.ScoreCutoffHelp(),
                           CommandLineParser::Integer);
 }
